完善商品详情页功能,完成结算购物车,添加订单表,订单详情,支付验证,库存验证,余额验证

+++ src/main/java/org/zyx/controller/ShopHandler.java
@@ -3,11 +3,13 @@ package org.zyx.controller;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.interceptor.TransactionAspectSupport;
 import org.springframework.web.bind.annotation.*;
 import org.zyx.VO.CartVO;
 import org.zyx.VO.DataVO;
 import org.zyx.entity.*;
 import org.zyx.enums.AddressStatus;
+import org.zyx.enums.ClearingStatus;
 import org.zyx.enums.ShopCarStatus;
 import org.zyx.mapper.*;
 import org.zyx.service.GoodCarService;
@@ -123,40 +125,69 @@ public class ShopHandler {
     //订单结算
     @Transactional
     @PostMapping("/clearing")
-    public boolean clearing(@RequestBody HashMap carIdList, HttpSession session) {
+    public int clearing(@RequestBody HashMap carIdList, HttpSession session) {
        // carIdList.forEach((k, v) -> System.out.println("key : " + k + " value : " + v));
         Buyer buyer = (Buyer) session.getAttribute("buyer");
         BigDecimal totalMoney = new BigDecimal(0);
 
+        //1.生成对应的订单表
+        OrderForm orderItem = new OrderForm();
+        orderItem.setOrderId((int)System.currentTimeMillis());//使用时间戳作为订单id,改为订单号;优先生成订单号
+        orderItem.setCreateTime(LocalDateTime.now());
+
+        //遍历结算购物车项
         Set set = carIdList.keySet();
         Iterator iterator = set.iterator();
         while(iterator.hasNext()){
             int carId = Integer.parseInt((String) iterator.next());//获取购物车id
             GoodCar goodItem = goodCarMapper.selectById(carId);//获取该项订单数据
             Goods good = goodsMapper.selectById(goodItem.getGoodsId());//获取商品信息
+
+            if(good.getStock() < goodItem.getNum()){
+                TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();//事务回滚
+                return ClearingStatus.NOT_ENOUGH_STOCK.getType();
+            }
+
             good.setStock(good.getStock()-goodItem.getNum());
             totalMoney = totalMoney.add(good.getPrice().multiply(new BigDecimal(goodItem.getNum())));//计算总金额
             goodsMapper.updateById(good);//更新商品信数量
             deleteById(carId);//移除购物车对应信息
+
+            //2.循环生成订单详情(以订单id添加)
+            OrderDetail orderDetail = new OrderDetail();
+            orderDetail.setGoodId(good.getGoodId());
+            orderDetail.setNum(goodItem.getNum());
+            orderDetail.setOrderId(orderItem.getOrderId());
+
+            orderDetailMapper.insert(orderDetail);//插入订单详情
+
         }
         //减少用户的余额
+        //先判断用户余额是否足够
+        if(buyer.getBalance().compareTo(totalMoney) < 0){
+            TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+            return ClearingStatus.NOT_ENOUGH_MONEY.getType();
+        }
+
         buyer.setBalance(buyer.getBalance().subtract(totalMoney));
         buyerMapper.updateBalance(buyer.getBuyerId(),buyer.getBalance());//更新余额
         session.setAttribute("buyer",buyer);//更新session中的buyer信息
 
-        //将该订单添加到订单表中,并生成对应的订单详情
-        OrderForm orderItem = new OrderForm();
-        orderItem.setCreateTime(LocalDateTime.now());
 
+        //设置订单相关信息
         QueryWrapper<Address> queryWrapper = new QueryWrapper<>();
         queryWrapper.eq("buyer_id",buyer.getBuyerId());
-        queryWrapper.eq("status", AddressStatus.START_USE.getType());
+        queryWrapper.eq("status", AddressStatus.START_USE.getType());//获取默认地址
         Address address = addressMapper.selectOne(queryWrapper);
         orderItem.setAddress(address.getAddress());//设置地址
+        orderItem.setPhone(address.getPhone());
+        orderItem.setNicheng(address.getNicheng());//
+        orderItem.setBuyerId(buyer.getBuyerId());
+        orderItem.setTotal(totalMoney);//订单总价
 
+        orderFormMapper.insert(orderItem);//添加订单表
 
-
-        return true;
+        return ClearingStatus.CLEARING_SUCCESS.getType();
     }
